apps/api/pyproject.toml
[project]
name = "rc7-programming-assistant-api"
version = "0.1.0"
description = "Backend FastAPI para RC7 Programming Assistant"
requires-python = ">=3.12"
dependencies = [
  "fastapi[standard]>=0.116.0,<1.0.0",
  "pwdlib[argon2]>=0.2.1,<1.0.0",
  "PyJWT>=2.10.1,<3.0.0",
  "pydantic-settings>=2.13.1,<3.0.0",
  "psycopg[binary]>=3.2.9,<4.0.0",
  "redis>=6.0.0,<7.0.0",
  "SQLAlchemy>=2.0.41,<3.0.0"
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0.0",
  "pytest>=8.4.1,<9.0.0"
]

[build-system]
requires = ["setuptools>=80.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.fastapi]
entrypoint = "src.main:app"

[tool.pytest.ini_options]
testpaths = ["tests"]
